fix: correctness bugs from the 2026-09-24 architecture review
- prefill: forward_prefill_pipelined's CED replay kept stale decoder rings
  for a restored suffix > 128 rows (KNOWN_BUGS #28 twin; cc47607 fixed only
  prefill_job_finish). Serial prefill_suffix / DSpark / non-multistream path.
- expertd ShardPool: a failed read (or "no evictable slot" mid-request) left
  claimed slots marked resident with no data and the layer's device remap
  stale; roll back uncommitted claims and force a remap re-upload.
- expertd prefetch: reader threads held a raw pointer into ExpertShard and
  wrote into B2Prefetch's pinned staging with nothing joining them (UAF on
  drop/move). Owner is now Arc'd; B2Prefetch::drop closes the hint channel
  and joins the readers.
- multistream: snapshot restores were accepted on token-id equality, so a
  same-size different image restored the old image's KV; verify bytes with
  image hashes like the serial path. Checkpoint saves that cut through an
  image are skipped instead of saved without span hashes.
- dsml: string parameters decoded every XML entity although only the
  defanged closing tag is ever escaped; a literal "&lt;" in tool args
  reached the client as "<".
- V4.1 prompt: tool results sorted into call order moved their images, but
  the handler paired pixels in message order (swapped pictures). The
  renderer now reports placeholder order and the handler follows it.
- multistream never trimmed the glibc heap (only the serial loop did);
  trim on the transition to idle.
- /readyz used env-or-60s while the watchdog honoured --hang-deadline-ms.
- lanes: >3 lanes silently aliased lane 2's sync events (now rejected);
  stage-graph key was a 16-bit lane hash overlapping b<<8 for b>=256
  (now an exact 64-bit key); route_probe clamped lanes to 1.

Checked: cargo check --tests (server, expertd, kernels; v41), server lib
tests (76+2 new), v41_prompt_vectors, tool_prompt_goldens, sampling_params.
No GPU tests run (server live); not deployed.

+++ crates/v4flash-kernels/src/het/graph_cache.rs
@@ -26,7 +26,7 @@ use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
 use v4flash_hip::{sys, GraphExec, Stream};
 
-pub type GraphKey = (&'static str, u32);
+pub type GraphKey = (&'static str, u64);
 
 #[derive(Default)]
 pub struct GraphCache {
@@ -50,12 +50,12 @@ impl GraphCache {
     /// On steady-state calls the mutex is held only long enough to
     /// `Arc::clone` the `GraphExec` out.
     /// The instantiated graph for `(stage, key)`, if captured already.
-    pub fn get(&self, stage: &'static str, key: u32) -> Option<Arc<GraphExec>> {
+    pub fn get(&self, stage: &'static str, key: u64) -> Option<Arc<GraphExec>> {
         self.entries.lock().unwrap().get(&(stage, key)).cloned()
     }
 
     /// Store a graph captured by the caller.
-    pub fn insert(&self, stage: &'static str, key: u32, exec: Arc<GraphExec>) {
+    pub fn insert(&self, stage: &'static str, key: u64, exec: Arc<GraphExec>) {
         self.entries.lock().unwrap().insert((stage, key), exec);
     }
 
@@ -73,7 +73,7 @@ impl GraphCache {
     where
         F: FnOnce(&Stream) -> eyre::Result<()>,
     {
-        let key: GraphKey = (stage, layer);
+        let key: GraphKey = (stage, layer as u64);
         let exec = {
             let mut entries = self.entries.lock().unwrap();
             if let Some(e) = entries.get(&key) {
